requirements.txt
google-generativeai>=0.6.0
python-dotenv>=1.0.1
playwright>=1.46.0
requests>=2.32.0
